#36 resolved.
The db migrations are now provided to the database initializer function as
a parameter, and the apps specific migrations are now part of the
main package.

There's no difference to the end-user, but this makes the code much neater
and more modular.
